+++ discordia/containers/snowflakes/channels/TextChannel.lua
@@ -19,7 +19,55 @@ function TextChannel:getMessageById(id)
 end
 
 function TextChannel:getMessages()
-	return self.messages:iter()
+	return self.messages:iterNewToOld()
+end
+
+function TextChannel:getMessageHistory(limit)
+	limit = math.clamp(limit or 50, 1, 100)
+	local success, data = self.client.api:getChannelMessages(self.id, limit)
+	if success then
+		local cache = OrderedCache({}, Message, 'id', math.huge, self)
+		for i = #data, 1, -1 do
+			cache:new(data[i])
+		end
+		return cache
+	end
+end
+
+function TextChannel:getMessageHistoryBefore(message, limit)
+	limit = math.clamp(limit or 50, 1, 100)
+	local success, data = self.client.api:getChannelMessages(self.id, limit, message.id, nil, nil)
+	if success then
+		local cache = OrderedCache({}, Message, 'id', math.huge, self)
+		for i = #data, 1, -1 do
+			cache:new(data[i])
+		end
+		return cache
+	end
+end
+
+function TextChannel:getMessageHistoryAfter(message, limit)
+	limit = math.clamp(limit or 50, 1, 100)
+	local success, data = self.client.api:getChannelMessages(self.id, limit, nil, message.id, nil)
+	if success then
+		local cache = OrderedCache({}, Message, 'id', math.huge, self)
+		for i = #data, 1, -1 do
+			cache:new(data[i])
+		end
+		return cache
+	end
+end
+
+function TextChannel:getMessageHistoryAround(message, limit)
+	limit = math.clamp(limit or 50, 2, 100) -- lower limit is 2 for this one
+	local success, data = self.client.api:getChannelMessages(self.id, limit, nil, nil, message.id)
+	if success then
+		local cache = OrderedCache({}, Message, 'id', math.huge, self)
+		for i = #data, 1, -1 do
+			cache:new(data[i])
+		end
+		return cache
+	end
 end
 
 return TextChannel